fix(react-utils): wrap provider & consumer to register them as client components
